Move `Endpoint` to Talpid package
