+++ android/src/main/kotlin/net/mullvad/mullvadvpn/model/TunnelEndpoint.kt
@@ -1,3 +1,5 @@
 package net.mullvad.mullvadvpn.model
 
+import net.mullvad.talpid.net.Endpoint
+
 data class TunnelEndpoint(val endpoint: Endpoint)
