+++ mullvad-jni/src/lib.rs
@@ -38,7 +38,6 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/mullvadvpn/model/AppVersionInfo",
     "net/mullvad/mullvadvpn/model/Constraint$Any",
     "net/mullvad/mullvadvpn/model/Constraint$Only",
-    "net/mullvad/mullvadvpn/model/Endpoint",
     "net/mullvad/mullvadvpn/model/GeoIpLocation",
     "net/mullvad/mullvadvpn/model/GetAccountDataResult$Ok",
     "net/mullvad/mullvadvpn/model/GetAccountDataResult$InvalidAccount",
@@ -71,6 +70,7 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/mullvadvpn/model/TunnelState$Disconnecting",
     "net/mullvad/mullvadvpn/MullvadDaemon",
     "net/mullvad/mullvadvpn/MullvadVpnService",
+    "net/mullvad/talpid/net/Endpoint",
     "net/mullvad/talpid/net/TransportProtocol$Tcp",
     "net/mullvad/talpid/net/TransportProtocol$Udp",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Block",
